+++ lisp/eab-helm.el
@@ -1,6 +1,6 @@
 ;;; eab-helm.el --- eab helm configure -*- lexical-binding: t -*-
 
-;; Copyright (C) 2010-2024 Evgeny Boykov
+;; Copyright (C) 2010-2025 Evgeny Boykov
 ;;
 ;; Author: artscan@list.ru
 ;; Keywords: