+++ lisp/eab-postload.el
@@ -1,6 +1,6 @@
 ;;; eab-postload.el ---  eab postload
 
-;; Copyright (C) 2010-2024 Evgeny Boykov
+;; Copyright (C) 2010-2025 Evgeny Boykov
 ;;
 ;; Author: artscan@list.ru
 ;; Keywords: 
@@ -31,7 +31,7 @@
 
 (when (eab/ondaemon (eab/server-C))
   ;; (eab/check-csum-day)
-  (setq system-time-locale "ru_RU.UTF-8"))
+  (setq system-time-locale "ru_RU.utf8"))
 
 (eab/bind-path eab/secrets-path)
 ; TODO create function and hook after first start frame
@@ -64,9 +64,9 @@
 
 (global-set-key (kbd "C-h c") 'describe-key-briefly)
 (global-set-key (kbd "M-O") 'forward-paragraph)
-(general-define-key
- :prefix "C-e"
- "d" docker-command-map)
+;; (general-define-key
+;;  :prefix "C-e"
+;;  "d" docker-command-map)
 (defvar eab/dired-map (make-sparse-keymap)
   "keymap for fast dired")
 (global-set-key (kbd "C-x d") nil)