+++ lisp/eab-ui.el
@@ -1,6 +1,6 @@
 ;;; eab-ui.el --- 
 
-;; Copyright (C) 2010-2024 Evgeny Boykov
+;; Copyright (C) 2010-2025 Evgeny Boykov
 ;;
 ;; Author: artscan@list.ru
 ;; Keywords: 
@@ -18,6 +18,37 @@
 (eab/bind-path custom-file)
 
 (setq find-function-C-source-directory source-directory)
+(setq recentf-max-saved-items 200)
+
+(defun eab/recentf-eabpool ()
+  (mapcar
+   (lambda (x)
+     (replace-regexp-in-string
+      (concat "^" (regexp-quote "~/")) ""
+      (replace-regexp-in-string
+       (concat "^" (regexp-quote "~/pnt/lion/")) "" x)))
+   (seq-filter
+    (lambda (x)
+      (file-exists-p x))
+    (seq-filter
+     (lambda (x)
+       (or
+        (let ((s "~/git/"))
+          (and (length> x (length s)) (string= (substring x 0 (length s)) s)))
+        (let ((s "~/pnt/lion/data/"))
+          (and (length> x (length s)) (string= (substring x 0 (length s)) s)))))
+     (seq-filter
+      (lambda (x)
+        (let ((s "~/git/org"))
+          (and (length> x (length s)) (not (string= (substring x 0 (length s)) s))))) recentf-list)))))
+
+(add-hook 'after-save-hook
+          (lambda ()
+            (f-write-text
+             (string-join
+              (eab/recentf-eabpool) "\n")
+             'utf-8
+             (concat recentf-save-file "-eabpool"))))
 
 (require 'autorevert)
 (global-auto-revert-mode)
@@ -32,14 +63,7 @@
 
 (require 'ansi-color)
 
-(defvar eab/recently-saved-files-list '())
-
 (setq ring-bell-function 'ignore)
-
-(add-hook 'after-save-hook
-          (lambda ()
-            (interactive)
-            (add-to-list 'eab/recently-saved-files-list buffer-file-name)))
 
 ;; see eab-compile.el
 (defun colorize-compilation-buffer ()
@@ -113,7 +137,7 @@
 (setq x-select-enable-clipboard t)
 (column-number-mode 1)
 
-(setq system-time-locale "ru_RU.UTF-8")
+(setq system-time-locale "ru_RU.utf8")
 
 (unless window-system ;; Only use in tty-sessions.
   (defvar arrow-keys-map (make-sparse-keymap) "Keymap for arrow keys")